udp: Don't pass proto to udp[46]_csum_init().
We passed IPPROTO_UDPLITE as proto to __udp[46]_lib_rcv(), which passes
it to udp[46]_csum_init().

However, we no longer call __udp[46]_lib_rcv() with IPPROTO_UDPLITE, so
proto is always IPPROTO_UDP in udp[46]_csum_init(), and we can hard-code
it.

Also, udp6_csum_init() is not called from other functions, so we move it
to net/ipv6/udp.c as a static function.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>