udp: Don't pass udp_table to __udp[46]_lib_lookup().
We use the same socket lookup function for UDP and UDP-Lite.

However, we managed UDP-Lite sockets in a global hash table and
UDP sockets in per-netns hash tables.

To use a proper table in __udp6_lib_lookup(), we needed to add an
argument for udp_table in many functions.

However, we no longer support UDP-Lite and need not pass udp_table
again and again.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>