tcp: Support [sg]etsockopt(TCP_EXT_DATA_OFFSET).
We can enable/disable EDO on a per-socket basis with setsockopt().

The attribute can be changed only before the connection is initiated,
meaning TCP_CLOSE for active open connection and TCP_LISTEN for passive
open one.

  $ python3
  >>> from socket import *
  >>>
  >>> TCP_EXT_DATA_OFFSET = 44
  >>> TCP_EDO_HDR = 1
  >>>
  >>> s = socket()
  >>> s.setsockopt(SOL_TCP, TCP_EXT_DATA_OFFSET, TCP_EDO_HDR)

The listener's attribute will be inherited to passive open connections.

Note that MD5 and TCP-AO cannot be used with EDO for now but will be
supported.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>