tcp: Send EDO Supported option in SYN.
EDO capability is determined in both directions using a single exchange
of the EDO Supported option.

  0              8              16             24             32
  +--------------+--------------+--------------+--------------+
  |  Kind (254)  |  Length (4)  |        ExID (0x0ed0)        |
  +--------------+--------------+--------------+--------------+

  TCP EDO Supported option with ExID

If tp->edo is set on a socket, we include the EDO Supported option in SYN.

Note that the EDO Supported option occupies 32 bits (fe04 0ed0 below) with
ExID for now.  When the official kind is assigned to EDO, it can be put in
a 2-byte NOP hole.

  IP 127.0.0.1.41066 > 127.0.0.1.40279: Flags [S], seq 173911591, win 33280,
  options [exp-0ed0,mss 65495,sackOK,TS val 1545532660 ecr 0,nop,wscale 7], length 0

	0x0000:  4500 0040 1966 4000 4006 2350 7f00 0001
	0x0010:  7f00 0001 a06a 9d57 0a5d ae27 0000 0000
	0x0020:  b002 8200 fe34 0000 fe04 0ed0 0204 ffd7
	0x0030:  0402 080a 5c1e f4f4 0000 0000 0103 0307

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>