tcp: Parse EDO Supported option in SYN.
Connections using EDO must negotiate its availability during the SYN
exchange.  While processing SYN, we have to update the receiver's EDO
attribute accordingly.

  SYN for listener : tcp_rsk(req)->edo
  Cross SYN        : tcp_sock(sk)->edo

If the EDO Supported option is seen in SYN and the receiver supports EDO,
the ->edo field continues to be 1; otherwise, the connection fall back to
non-EDO mode, so ->edo is set to 0, and LINUX_MIB_TCPEDOFALLBACKSYN is
incremented.

SYN Cookie with EDO is not supported but will be supported with BPF.

As of this commit, in the cross SYN case, SYN+ACK will include EDO
Supported.  SYN+ACK is sent by tcp_send_synack(), which appends
TCPHDR_ACK to TCP_SKB_CB(skb)->tcp_flags, but originally the skb
has TCPHDR_SYN.  So, which options are included is determined in
tcp_syn_options(), and SYN+ACK includes EDO Supported option.

  # python3
  >>> from socket import *
  >>>
  >>> s = socket()
  >>> s.bind(('0', 0))
  >>> s.connect(s.getsockname())

  Both SYN and SYN+ACK have fe04 0ed0 :

  IP 127.0.0.1.54735 > 127.0.0.1.54735: Flags [S], seq 2489470833, win 33280,
  options [exp-0ed0,mss 65495,sackOK,TS val 12616497 ecr 0,nop,wscale 7], length 0

	0x0000:  4500 0040 6b9a 4000 4006 d11b 7f00 0001
	0x0010:  7f00 0001 d5cf d5cf 9462 4f71 0000 0000
	0x0020:  b002 8200 fe34 0000 fe04 0ed0 0204 ffd7
	0x0030:  0402 080a 00c0 8331 0000 0000 0103 0307

  IP 127.0.0.1.54735 > 127.0.0.1.54735: Flags [S.], seq 2489470833, ack 2489470834, win 33280,
  options [exp-0ed0,mss 65495,sackOK,TS val 12616499 ecr 12616497,nop,wscale 7], length 0

	0x0000:  4500 0040 6b9b 4000 4006 d11a 7f00 0001
	0x0010:  7f00 0001 d5cf d5cf 9462 4f71 9462 4f72
	0x0020:  b012 8200 fe34 0000 fe04 0ed0 0204 ffd7
	0x0030:  0402 080a 00c0 8333 00c0 8331 0103 0307

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>