tcp: Parse EDO Extension option in the final ACK of 3WHS.
In tcp_check_req(), we turn on EDO Extension parsing based on
tcp_rsk(req)->edo and update it with tmp_opt.edo_ok, which is
set to 1 if EDO Extension is seen in the final ACK of 3WHS.

Later in tcp_create_openreq_child(), tcp_rsk(req)->edo is inherited
to full sk.

If EDO Extension has not been sent in the final ACK, EDO will be
disabled on the connection.

In SYN cookie case, we do not send EDO Supported option in SYN+ACK,
so we do not parse EDO Extension in the following segments.  So, we
set 0 to tcp_rsk(req)->edo in cookie_v[46]_check().

Another notable thing is we must update tp->edo in tcp_check_req()
for TCP Fast Open case.

When receiving TFO SYN, we create a full sk.  Then, tp->edo is
inherited from tcp_rsk(req)->edo even before negotiation completes.

SYN+ACK is created with reqsk, so EDO Supported is sent properly.

When the final ACK is received, we parse it in tcp_check_req()
and tcp_validate_incoming().  The former triggers EDO Extension
parsing based on reqsk and can falls back to non-EDO.  However,
the latter will rely on tp->edo and drop packets without EDO
Extension option if tp->edo is 1.

But, the ACK of 3WHS is allowed not to have EDO Extension option,
and we must do fallback then.

Therefore, we have to update tp->edo in tcp_check_req() for TFO
so that tcp_validate_incoming() will not drop ACK without EDO
Extension option.

Also note that we need to reload tcphdr and iphdr in these 3
callers of tcp_check_req().

  * tcp_rcv_state_process
  * tcp_v4_rcv
  * tcp_v6_rcv

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>