tcp: Drop const from skb in tcp_parse_options().
This is a preparation patch to support Extended Data Offset (EDO).

When processing incoming TCP packets, we pull (th->doff * 4) in
tcp_v4_rcv().

However, this is not enough if we receive EDO Extension option.
Then, we need to call pskb_may_pull() for the extra header.

We need to drop const from skb to call pskb_may_pull().

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>