tcp: Prepare for EDO in tcp_parse_options().
We will add parser for EDO Extension option.

When processing incoming TCP packets, we pull (th->doff * 4) bytes
into the linear area of skb in tcp_v4_rcv().

However, this might not be enough in case of EDO, and we could end
up calling pskb_may_pull() for the extra options.

So, we need to drop const from skb to call pskb_may_pull() and return
an error if it fails.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>