tcp: Send EDO Extension option in tcp_v4_send_ack().
When sending ACK in FIN_WAIT2 or TIME_WAIT, we call tcp_v4_send_ack().

If EDO is negotiated on the connection, we include EDO Extension
option in ACK generated in tcp_v4_send_ack().

We also call it for reqsk when sending Challenge ACK.  Then, EDO has
not been negotiated yet, so we don't include EDO Extension option.

Note EDO Extension option is not placed as the first option not to
slow down Timestamp option generation.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>