tcp: Add parser for EDO Extension option.
We add a new argument parse_edo_ext in tcp_parse_options() to control
whether EDO Extension option is parsed or not.  Now it is disabled in
all paths and never fails.  We will turn it on later for each path.

If a segment has EDO Extension option, we call pskb_may_pull() and
update some values used in tcp_parse_options().  Then, we also need
to update TCP_SKB_CB(skb)->end_seq.

TCP_SKB_CB(skb)->end_seq is filled in tcp_v[46]_fill_cb() as

  TCP_SKB_CB(skb)->seq + th->syn + th->fin + skb->len - th->doff * 4

and set to tp->rcv_nxt in tcp_rcv_nxt_update() to use the value as
ACK number.

If we forgot to update end_seq, the extended area placed after 60 bytes
would be treated as user data and a wrong ACK would be sent for it.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>