af_unix: Add struct unix_vertex in struct unix_sock.
We will replace the garbage collection algorithm for AF_UNIX,
where we will consider each inflight file descriptor of AF_UNIX
sockets as an edge in a directed graph.

Here, we introduce a new struct unix_vertex representing a vertex
in the graph and add it to struct unix_sock.

In the following patch, we will allocate another struct per edge,
which we finally link to the inflight socket's unix_vertex.edges
when sendmsg() succeeds.

Also, we will count the number of edges as unix_vertex.out_degree.

The first time an AF_UNIX socket is passed successfully, we will
link its unix_vertex.entry to a global list.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>