af_unix: Use skb_unref() for OOB skb in unix_collect_skb().
When we queuing OOB skb, we hold an additional refcount in queue_oob().

While garbage-collecting a socket with OOB skb, we use kfree_skb() to
drop the count.

It can be safely called there, but recently syzkaller reported a deadlock
due to a bug in manage_oob().

In case a similar issue occurs again, let's use skb_unref() to reduce
the impact.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>