rtnetlink: Allocate linkinfo instead of slave_attr as struct rtnl_newlink_tbs.
rtnl_newlink() allocates struct rtnl_newlink_tbs and pass it to
__rtnl_newlink(), which has 3 functionalities; creating a link,
configuring a link, configuring a group of links.

slave_attr is only used by the 2nd one, and we will factorise
the part in the following patch.  Then, slave_attr is not used
in __rtnl_newlink().

As a prep, let's keep slave_attr on stack and instead allocate
linkinfo within struct rtnl_newlink_tbs.

Note that RTNL_SLAVE_MAX_TYPE is 44 and IFLA_INFO_MAX is 6.

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>