wifi: wext: Don't iterate all netns in wireless_nlevent_flush().
Commit 8bf862739a77 ("wext: fix message delay/ordering") introduced
wext_netdev_notifier_call() to avoid message delay and out-of-order.

The notifier calls wireless_nlevent_flush(), which iterates all netns.

The problem is that this happens for any event and netdev, even if
the host does not have a wext device.

This is too noisy, especially on a host with thousands of netns as
reported by Alexandre Ferrieux.

Given that the wext event queue was implemented in struct net,
wireless_nlevent_flush() does not need to iterate all netns.

Let's avoid unnecessary netns iteration in wireless_nlevent_flush().

Reported-by: Alexandre Ferrieux <alexandre.ferrieux@gmail.com>
Closes: https://lore.kernel.org/netdev/CAKYWH0Ti3=4GeeuVyWKJ9LyTuRnf3Wy9GKg4Jb7tdeaT39qADA@mail.gmail.com/
Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>