socket: Pass netref flag to __sock_create().
We will introduce a new API to create a kernel socket with netns
refcnt held.

As a prep, let's add netref argument to __sock_create().

Note that we still do not pass it down to pf->create() for clarity;
otherwise, this change is buried in the huge diff.

Another option would be to override the kern parameter, which is int.
But, I chose this approach to make sure, with the help of the compiler,
that all paths pass parameters down to sk_alloc() as is, and there
actually was a weird path in smc_ulp_init().

Signed-off-by: Kuniyuki Iwashima <kuniyu@amazon.com>